+++ nekrosis/support/windows_utilities/persistence_methods.py
@@ -10,4 +10,6 @@ class WindowsPersistenceMethods(enum.Enum):
     Windows-specific persistence methods.
     """
 
-    REGEDIT_RUN: str = "Regedit Run Key"
+    REGEDIT_RUN:          str = "Regedit Run Key"
+    STARTUP_CURRENT_USER: str = "Startup Folder (Current User)"
+    STARTUP_GLOBAL:       str = "Startup Folder (Global)"